Deal with the groupSet element that appears in the instanceSet element in the DescribeInstances response.
In the 2011-01-01 ec2 API (which this commit also switches back to), there's a groupSet element within the instancesSet element. The groupSet element is not documented in http://docs.amazonwebservices.com/AWSEC2/2011-01-01/APIReference/, but it does appear in the examples on that page, and it is sent by Amazon. Without this change, the closing </item> tag within the groupSet tag within the instancesSet tag was interpreted by boto.handler.XmlHandler as the close item tag for the instance, which caused the returned instance not to have any of the data that appears later in the XML element, including the tags. This tag causes the SML processing to recognize the embedded groupSet tag and continue.