Fixes standard boto Exception sub-classes to pass *args.
This is mentioned only briefly and vaguely in the Python docs:

"If the exception class is derived from the standard root class
BaseException, the associated value is present as the exception
instance’s args attribute."

The basic gist is that if you override Exception.__init__, you
need to call super().__init__(*args). Failing to do this causes
some issues that aren't seen very often, but one of the most
bothersome things is that the exceptions are un-pickleable. If
you happen to be using something like celery, you get really
vague error notifications.

This commit makes our base exception classes "correct", which
should help prevent such annoyances for those of us who have
to pickle exceptions for some reason or another.

This should be 100% backwards compatible.