Make cacerts.txt package data instead of just including it in source distributions.
The cacerts.txt was missing from boto distributions (see issue #260).
Adding it to MANIFEST.in causes the file to be included in sdist source
archives, but *not* when installing via PyPI/pip/directly from github.

To fix this issue, MANIFEST.in has been deleted and the (hopefully, as
the whole setuptools/distutils affair is a bit of a mess) proper options
have been added to setup.py.

Installation via git+https through pip has been tested, I have also
checked if the file is included in the tarball generated by sdist. I am
reasonably confident that this will also cause installs from PyPI to
work afterwards.