Adding incremental MD5 computation.
This greatly reduces the wall clock time for an upload when uploading large files.

Situations in which the full MD5 is still calculated in its entirety
  - When resuming a partial upload
  - When a name is not specified for the object, because GCS uses the MD5 as the name

Updating one test in ResumableUploadTests to change a byte that has already been uploaded, so a failure will occur.
Otherwise the on-the-fly evaluation of the md5 will not cause an error, as it will be computed after mutation
has occurred.