Work around Amazon ListParts RangeInBytes bug
Amazon's ListParts implementation returns off-by-ones in parts'
RangeInBytes fields, contrary to their documentation.

Reported here:
https://forums.aws.amazon.com/thread.jspa?threadID=106866&tstart=0

For now, a workaround is to fix the result up by hand in the common case
of a full (non-last) part, since we know that this particular off-by-one
is impossible. If we end up with the wrong RangeInBytes in this
particular case, then the worst that will happen is that the upload
resume functionality will mismatch on the hash and re-upload the part.